Big edits again mostly to reintegrated AHB technical content
* Cleared up some confusion about app display. It is one thing to
display “information about an app” so that the user can select it. It
is another to display a browser-based app. To help clear this up, I
changed the term from “displayed app” to “selectable app”.

* Added back some technical content but:
a) Reduced redundancy
b) Clarified steps as different sections
c) Made clear that choices such as HTTP / start_url are options under
consideration (with others like JS).