[Privacy] Initial documentation on privacy implications for web payments
This PR contains a set of README files covering some thoughts around the impact
of web privacy changes on payments. It is organized into three sections so far:

1. A problem statement
2. A set of known payment flows for which we are considering the impact
3. A set of known cases where third-party cookie deprecation will affect a
   payments-related use-case.

None of the work should be considered either complete or accurate; it is the
best understanding of the authors of this PR and should definitely be evolved
on over time!

Also-by: Nick Burris <nburris@chromium.org>